Added guid to users table after login and fixed import issues
